+++ src/oci/data_safe/models/data_safe_private_endpoint.py
@@ -104,6 +104,10 @@ class DataSafePrivateEndpoint(object):
             The value to assign to the defined_tags property of this DataSafePrivateEndpoint.
         :type defined_tags: dict(str, dict(str, object))
 
+        :param system_tags:
+            The value to assign to the system_tags property of this DataSafePrivateEndpoint.
+        :type system_tags: dict(str, dict(str, object))
+
         """
         self.swagger_types = {
             'id': 'str',
@@ -119,7 +123,8 @@ class DataSafePrivateEndpoint(object):
             'lifecycle_state': 'str',
             'nsg_ids': 'list[str]',
             'freeform_tags': 'dict(str, str)',
-            'defined_tags': 'dict(str, dict(str, object))'
+            'defined_tags': 'dict(str, dict(str, object))',
+            'system_tags': 'dict(str, dict(str, object))'
         }
 
         self.attribute_map = {
@@ -136,7 +141,8 @@ class DataSafePrivateEndpoint(object):
             'lifecycle_state': 'lifecycleState',
             'nsg_ids': 'nsgIds',
             'freeform_tags': 'freeformTags',
-            'defined_tags': 'definedTags'
+            'defined_tags': 'definedTags',
+            'system_tags': 'systemTags'
         }
 
         self._id = None
@@ -153,6 +159,7 @@ class DataSafePrivateEndpoint(object):
         self._nsg_ids = None
         self._freeform_tags = None
         self._defined_tags = None
+        self._system_tags = None
 
     @property
     def id(self):
@@ -515,6 +522,32 @@ class DataSafePrivateEndpoint(object):
         :type: dict(str, dict(str, object))
         """
         self._defined_tags = defined_tags
+
+    @property
+    def system_tags(self):
+        """
+        Gets the system_tags of this DataSafePrivateEndpoint.
+        System tags for this resource. Each key is predefined and scoped to a namespace. For more information, see Resource Tags.
+        Example: `{\"orcl-cloud\": {\"free-tier-retained\": \"true\"}}`
+
+
+        :return: The system_tags of this DataSafePrivateEndpoint.
+        :rtype: dict(str, dict(str, object))
+        """
+        return self._system_tags
+
+    @system_tags.setter
+    def system_tags(self, system_tags):
+        """
+        Sets the system_tags of this DataSafePrivateEndpoint.
+        System tags for this resource. Each key is predefined and scoped to a namespace. For more information, see Resource Tags.
+        Example: `{\"orcl-cloud\": {\"free-tier-retained\": \"true\"}}`
+
+
+        :param system_tags: The system_tags of this DataSafePrivateEndpoint.
+        :type: dict(str, dict(str, object))
+        """
+        self._system_tags = system_tags
 
     def __repr__(self):
         return formatted_flat_dict(self)