+++ src/oci/management_dashboard/models/create_management_dashboard_details.py
@@ -90,6 +90,10 @@ class CreateManagementDashboardDetails(object):
             The value to assign to the is_favorite property of this CreateManagementDashboardDetails.
         :type is_favorite: bool
 
+        :param parameters_config:
+            The value to assign to the parameters_config property of this CreateManagementDashboardDetails.
+        :type parameters_config: list[object]
+
         :param freeform_tags:
             The value to assign to the freeform_tags property of this CreateManagementDashboardDetails.
         :type freeform_tags: dict(str, str)
@@ -118,6 +122,7 @@ class CreateManagementDashboardDetails(object):
             'data_config': 'list[object]',
             'type': 'str',
             'is_favorite': 'bool',
+            'parameters_config': 'list[object]',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
         }
@@ -141,6 +146,7 @@ class CreateManagementDashboardDetails(object):
             'data_config': 'dataConfig',
             'type': 'type',
             'is_favorite': 'isFavorite',
+            'parameters_config': 'parametersConfig',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
         }
@@ -163,6 +169,7 @@ class CreateManagementDashboardDetails(object):
         self._data_config = None
         self._type = None
         self._is_favorite = None
+        self._parameters_config = None
         self._freeform_tags = None
         self._defined_tags = None
 
@@ -597,6 +604,30 @@ class CreateManagementDashboardDetails(object):
         :type: bool
         """
         self._is_favorite = is_favorite
+
+    @property
+    def parameters_config(self):
+        """
+        Gets the parameters_config of this CreateManagementDashboardDetails.
+        Defines parameters for the dashboard.
+
+
+        :return: The parameters_config of this CreateManagementDashboardDetails.
+        :rtype: list[object]
+        """
+        return self._parameters_config
+
+    @parameters_config.setter
+    def parameters_config(self, parameters_config):
+        """
+        Sets the parameters_config of this CreateManagementDashboardDetails.
+        Defines parameters for the dashboard.
+
+
+        :param parameters_config: The parameters_config of this CreateManagementDashboardDetails.
+        :type: list[object]
+        """
+        self._parameters_config = parameters_config
 
     @property
     def freeform_tags(self):