+++ src/oci/management_dashboard/models/update_management_saved_search_details.py
@@ -95,6 +95,10 @@ class UpdateManagementSavedSearchDetails(object):
             The value to assign to the widget_vm property of this UpdateManagementSavedSearchDetails.
         :type widget_vm: str
 
+        :param parameters_config:
+            The value to assign to the parameters_config property of this UpdateManagementSavedSearchDetails.
+        :type parameters_config: list[object]
+
         :param freeform_tags:
             The value to assign to the freeform_tags property of this UpdateManagementSavedSearchDetails.
         :type freeform_tags: dict(str, str)
@@ -120,6 +124,7 @@ class UpdateManagementSavedSearchDetails(object):
             'metadata_version': 'str',
             'widget_template': 'str',
             'widget_vm': 'str',
+            'parameters_config': 'list[object]',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
         }
@@ -140,6 +145,7 @@ class UpdateManagementSavedSearchDetails(object):
             'metadata_version': 'metadataVersion',
             'widget_template': 'widgetTemplate',
             'widget_vm': 'widgetVM',
+            'parameters_config': 'parametersConfig',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
         }
@@ -159,6 +165,7 @@ class UpdateManagementSavedSearchDetails(object):
         self._metadata_version = None
         self._widget_template = None
         self._widget_vm = None
+        self._parameters_config = None
         self._freeform_tags = None
         self._defined_tags = None
 
@@ -529,6 +536,30 @@ class UpdateManagementSavedSearchDetails(object):
         :type: str
         """
         self._widget_vm = widget_vm
+
+    @property
+    def parameters_config(self):
+        """
+        Gets the parameters_config of this UpdateManagementSavedSearchDetails.
+        Defines parameters for the saved search.
+
+
+        :return: The parameters_config of this UpdateManagementSavedSearchDetails.
+        :rtype: list[object]
+        """
+        return self._parameters_config
+
+    @parameters_config.setter
+    def parameters_config(self, parameters_config):
+        """
+        Sets the parameters_config of this UpdateManagementSavedSearchDetails.
+        Defines parameters for the saved search.
+
+
+        :param parameters_config: The parameters_config of this UpdateManagementSavedSearchDetails.
+        :type: list[object]
+        """
+        self._parameters_config = parameters_config
 
     @property
     def freeform_tags(self):