+++ CHANGELOG.rst
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+2.40.0 - 2021-06-15
+====================
+
+Added
+-----
+* Support for elastic storage on Exadata Infrastructure resources for Cloud at Customer in the Database service
+* Support for registration and management of target databases in the Data Safe service
+* Support for config on metadata in the Management Dashboard service
+* Support for a new work request operation type for node pool reconciliation events in the Container Engine for Kubernetes service
+* Support for migrating clusters with a public Kubernetes API endpoint which are not integrated with a customer's VCN to a VCN-native cluster in the Container Engine for Kubernetes service
+* Support for getting the spark version of applications, and filtering applications by spark version, in the Data Flow service
+
+Breaking
+--------
+* The properties `freeform_tags` and `defined_tags` were removed from the ManagementDashboardExportDetails model in the Management Dashboard service
+
+====================
 2.39.1 - 2021-06-08
 ====================